requirements.txt
appdirs==1.4.3
asgiref==3.2.7
certifi==2019.11.28
chardet==3.0.4
distlib==0.3.0
Django==3.0.3
djangorestframework==3.11.0
et-xmlfile==1.0.1
filelock==3.0.12
idna==2.9
jdcal==1.4.1
openpyxl==3.0.3
pipenv==2018.11.26
psycopg2-binary==2.8.4
pytz==2019.3
requests==2.23.0
six==1.14.0
sqlparse==0.3.1
urllib3==1.25.8
virtualenv==20.0.15
virtualenv-clone==0.5.4
django-heroku
gunicorn
